requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
numpy>=1.21.0
google-generativeai
google-cloud-speech
google-cloud-texttospeech
sounddevice
Pillow>=9.0.0
streamlit-drawable-canvas>=0.9.0
firebase-admin
google-auth
google-auth-oauthlib
streamlit-authenticator
